implement sequence slice with positive indexes

+++ boa3_test/tests/test_function.py
@@ -701,7 +701,7 @@ class TestFunction(BoaTest):
                 + Integer(5).to_byte_array(min_length=1)
                 + Opcode.APPEND
                 + Opcode.JMP
-                + Integer(2).to_byte_array(min_length=1)
+                + Integer(3).to_byte_array(min_length=1)
                 + Opcode.STLOC0
             + Opcode.LDLOC1                 # i += 1
             + Opcode.PUSH1
